Add crouching and VIEWHEIGHT setting
P_CalcHeight now replaces VIEWHEIGHT with a variable, which gets a value assigned based on the chosen VIEWHEIGHT setting (original or adjusted) and on the player crouching or standing, and uses that variable for view bobbing calculations. Additionally, if the player just stood up after crouching, the "smooth step up" code is executed, resulting in a smooth crouch-to-stand transition.

P_MovePlayer's P_Thrust calls replace the fixed value of 2048 with a variable, which starts at 2048 and gets halved if the player is crouching, and uses that variable for movement calculations.