'Misc. Bug Fixes' setting and P_NuggetCheckDist
Renamed 'P_CheckSSGGibbing' to 'P_NuggetCheckDist', and changed its functionality to add flexibility. As part of this, its usage in the SSG gibbing check has also been altered.

A new check was added to P_DamageMobj when applying thrust: if the player has the Chainsaw equipped, the distance between source and target is greater than the CSaw range, 'Misc. Bug Fixes' is enabled and no demo play is going on, do apply thrust.